Foundation Classes - Fix const Array1 CI issues
Avoid duplicate NCollection_Allocator::address() overloads for const value types and allow const-pointer deallocation through the allocator.

Use int Array1 indices in ToArray1 read-through tests to avoid -Wshorten-64-to-32 failures.

+++ src/FoundationClasses/TKernel/GTests/NCollection_LinearVector_Test.cxx
@@ -885,9 +885,9 @@ TEST(NCollection_LinearVectorTest, ToArray1_ReadThrough)
   }
 
   const NCollection_Array1<int> aArr = aVec.ToArray1();
-  for (size_t i = 0; i < 7; ++i)
+  for (int i = 0; i < 7; ++i)
   {
-    EXPECT_EQ(static_cast<int>(i * 11), aArr[i]);
+    EXPECT_EQ(i * 11, aArr[i]);
   }
 }
