+++ src/FoundationClasses/TKernel/GTests/NCollection_LocalArray_Test.cxx
@@ -701,8 +701,8 @@ TEST(NCollection_LocalArrayTest, ToArray1_ReadThrough)
   }
 
   const NCollection_Array1<int> aArr = aSrc.ToArray1();
-  for (size_t i = 0; i < 7; ++i)
+  for (int i = 0; i < 7; ++i)
   {
-    EXPECT_EQ(static_cast<int>(i * 11), aArr[i]);
+    EXPECT_EQ(i * 11, aArr[i]);
   }
 }
